Refactorizando y quitando warnings
Ha pasado de haber 1700 y pico a 474

+++ MonkeyDelivery/Src/Logic/Enemy.h
@@ -26,7 +26,7 @@ protected:
 	bool startTimer = false;
 	AnimationManager* animationManager;
 	SDL_Rect textureRect;
-	int timerAnimation;
+	float timerAnimation;
 
 
 	//movimiento enemigos respecto a camara
